requirements.txt
Django==1.7
django-taggit
textile
django-markup
pytz
unipath


